osa5: 5.7

+++ osa5/backend/src/controllers/blogs.js
@@ -3,6 +3,11 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+const verifyToken = (token) => {
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  return token && decodedToken.id
+}
+
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
@@ -28,10 +33,8 @@ blogsRouter.post('/', async (request, response, next) => {
 
   try {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-    if (!request.token || !decodedToken.id) {
-      return response.status(401).json({ error: 'token missing or invalid' })
-    }
+    if (!request.token || !decodedToken.id)
+      return response.status(401).json({ error: 'invalid token' })
 
     const user = await User.findById(decodedToken.id)
 
@@ -44,8 +47,8 @@ blogsRouter.post('/', async (request, response, next) => {
     })
 
     const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id) //highlight-line
-    await user.save()  //highlight-line
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.json(savedBlog.toJSON())
   } catch(exception) {
     next(exception)
@@ -54,6 +57,9 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    if (!verifyToken(request.token))
+      return response.status(401).json({ error: 'invalid token' })
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
@@ -61,21 +67,38 @@ blogsRouter.delete('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
 
-  const blog = {
+  const filter = {
+    _id: request.params.id
+  }
+
+  const updatedBlog = {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: body.user,
   }
 
-  Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    .then(updatedBlog => {
-      response.json(updatedBlog.toJSON())
-    })
-    .catch(error => next(error))
+  const options = {
+    new: true
+  }
+
+  try {
+    if (!verifyToken(request.token)) {
+      return response.status(401).json({ error: 'invalid token' })
+    }
+
+    const result = await Blog
+      .findOneAndUpdate(filter, updatedBlog, options)
+      .populate('user', { username: 1, name: 1})
+    response.json(result.toJSON())
+  } catch (exception) {
+    console.log(exception)
+    next(exception)
+  }
 })
 
 module.exports = blogsRouter
